--- core/core-api/src/main/java/com/ticket/core/domain/hold/application/HoldKeyGenerator.java
@@ -1,13 +1,0 @@
-package com.ticket.core.domain.hold.application;
-
-import org.springframework.stereotype.Component;
-
-import java.util.UUID;
-
-@Component
-public class HoldKeyGenerator {
-
-    public String generate() {
-        return "HOLD-" + UUID.randomUUID().toString().replace("-", "");
-    }
-}
