--- core/core-api/src/main/java/com/ticket/core/domain/hold/support/HoldReleaseLockException.java
@@ -1,12 +1,0 @@
-package com.ticket.core.domain.hold.support;
-
-public class HoldReleaseLockException extends RuntimeException {
-
-    public HoldReleaseLockException(final String message) {
-        super(message);
-    }
-
-    public HoldReleaseLockException(final String message, final Throwable cause) {
-        super(message, cause);
-    }
-}
