+++ core/core-domain/src/main/java/com/ticket/core/domain/hold/command/HoldKeyGenerator.java
@@ -1,0 +1,13 @@
+package com.ticket.core.domain.hold.command;
+
+import org.springframework.stereotype.Component;
+
+import java.util.UUID;
+
+@Component
+public class HoldKeyGenerator {
+
+    public String generate() {
+        return "HOLD-" + UUID.randomUUID().toString().replace("-", "");
+    }
+}
