+++ core/core-domain/src/main/java/com/ticket/core/domain/hold/event/HoldCreatedEvent.java
