+++ core/core-domain/src/main/java/com/ticket/core/domain/hold/model/HoldSnapshot.java
