+++ core/core-domain/src/test/java/com/ticket/core/domain/hold/model/HoldSnapshotTest.java
